refactor: folder structure for board, issue & demo
