+++ client/src/App.js
@@ -1,9 +1,9 @@
 import './App.css'
 import { Route, Routes } from 'react-router-dom'
-import Board from './components/Board.js'
-import Demo from './components/Demo.js'
+import Board from './components/board/Board.js'
+import Demo from './components/demo/Demo.js'
 import Home from './components/Home.js'
-import Issue from './components/Issue.js'
+import Issue from './components/issue/Issue.js'
 import Oauth from './components/oauth/Oauth.js'
 import Authorization from './components/oauth/Authorization.js'
