+++ client/src/components/demo/Workspace.js
@@ -1,8 +1,8 @@
 import React from 'react'
-import RadioList from './RadioList.js'
-import Button from './Button.js'
-import { useUsersMe } from '../hooks/asana/useUsersMe.js'
-import { useRadio } from '../reducers/useRadio.js'
+import RadioList from '../RadioList.js'
+import Button from '../Button.js'
+import { useUsersMe } from '../../hooks/asana/useUsersMe.js'
+import { useRadio } from '../../reducers/useRadio.js'
 
 function Workspace({ updateWorkspaceGid }) {
 	const { isFetching, workspaces, meGid } = useUsersMe()
